django-backend/pyproject.toml
[tool.poetry]
name = "django-backend"
version = "0.1.0"
description = ""
authors = ["Avin <avinkavish@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.6"
psycopg2 = "^2.9.9"
djangorestframework = "^3.15.1"
markdown = "^3.6"
django-filter = "^24.2"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.1"}
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
